+++ src/main/java/by/bsu/rfct/fclearn/dao/CollectionDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface CollectionDAO extends GenericDAO<Collection, Long>{
 
-    List<Collection> readAllByTopicId(Long topicId, Long startLimitFrom, Long amountOnPage);
+    List<Collection> readAllByTopicId(Long topicId, Integer startLimitFrom, Integer amountOnPage);
 
     Long countByAuthorId(Long authorId);
