+++ src/main/java/by/bsu/rfct/fclearn/service/CollectionService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface CollectionService extends GenericService<CollectionDTO, Long> {
 
-    List<CollectionDTO> readAllByTopicId(Long topicId, Long pageNumber, Long amountOnPage);
+    List<CollectionDTO> readAllByTopicId(Long topicId, Integer pageNumber, Integer amountOnPage);
 
     Long countByAuthorId(Long authorId);
