+++ src/main/java/by/bsu/rfct/fclearn/service/util/ServiceUtils.java
@@ -2,7 +2,7 @@ package by.bsu.rfct.fclearn.service.util;
 
 public class ServiceUtils {
 
-    public static Long countStartLimitFrom(Long pageNumber, Long amountOnPage) {
+    public static Integer countStartLimitFrom(Integer pageNumber, Integer amountOnPage) {
         return (pageNumber - 1) * amountOnPage;
     }
 }
